chore: extract methods for agent deployment to its own file
