+++ e2e/_suites/fleet/fleet.go
@@ -11,7 +11,6 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strings"
 	"time"
 
@@ -71,63 +70,6 @@ func (fts *FleetTestSuite) getDeployer() deploy.Deployment {
 	return fts.deployer
 }
 
-// this step infers the installer type from the underlying OS image
-// supported images: centos and debian
-func (fts *FleetTestSuite) anAgentIsDeployedToFleet(image string) error {
-	installerType := "rpm"
-	if image == "debian" {
-		installerType = "deb"
-	}
-	fts.BeatsProcess = ""
-
-	// FIXME: We need to cleanup the steps to support different operating systems
-	// for now we will force the zip installer type when the agent is running on windows
-	if runtime.GOOS == "windows" && common.Provider == "remote" {
-		installerType = "zip"
-	}
-	return fts.anAgentIsDeployedToFleetWithInstallerAndFleetServer(installerType)
-}
-
-func (fts *FleetTestSuite) anAgentIsDeployedToFleetOnTopOfBeat(beatsProcess string) error {
-	installerType := "tar"
-
-	// FIXME: We need to cleanup the steps to support different operating systems
-	// for now we will force the zip installer type when the agent is running on windows
-	if runtime.GOOS == "windows" && common.Provider == "remote" {
-		installerType = "zip"
-	}
-
-	fts.BeatsProcess = beatsProcess
-
-	return fts.anAgentIsDeployedToFleetWithInstallerAndFleetServer(installerType)
-}
-
-// supported installers: tar, rpm, deb
-func (fts *FleetTestSuite) anAgentIsDeployedToFleetWithInstaller(installerType string) error {
-	fts.BeatsProcess = ""
-
-	// FIXME: We need to cleanup the steps to support different operating systems
-	// for now we will force the zip installer type when the agent is running on windows
-	if runtime.GOOS == "windows" && common.Provider == "remote" {
-		installerType = "zip"
-	}
-
-	return fts.anAgentIsDeployedToFleetWithInstallerAndFleetServer(installerType)
-}
-
-// supported installers: tar, rpm, deb
-func (fts *FleetTestSuite) anAgentIsDeployedToFleetWithInstallerAndTags(installerType string, flags string) error {
-	fts.BeatsProcess = ""
-
-	// FIXME: We need to cleanup the steps to support different operating systems
-	// for now we will force the zip installer type when the agent is running on windows
-	if runtime.GOOS == "windows" && common.Provider == "remote" {
-		installerType = "zip"
-	}
-	fts.ElasticAgentFlags = flags
-	return fts.anAgentIsDeployedToFleetWithInstallerAndFleetServer(installerType)
-}
-
 func (fts *FleetTestSuite) tagsAreInTheElasticAgentIndex() error {
 	var tagsArray []string
 	//ex of flags  "--tag production,linux" or "--tag=production,linux"
@@ -165,40 +107,6 @@ func (fts *FleetTestSuite) tagsAreInTheElasticAgentIndex() error {
 		log.WithFields(log.Fields{
 			"error": err,
 		}).Warn(elasticsearch.WaitForIndices())
-	}
-	return err
-}
-
-func (fts *FleetTestSuite) anAgentIsDeployedToFleetWithInstallerAndFleetServer(installerType string) error {
-	log.WithFields(log.Fields{
-		"installer": installerType,
-	}).Trace("Deploying an agent to Fleet with base image using an already bootstrapped Fleet Server")
-
-	deployedAgentsCount++
-
-	fts.InstallerType = installerType
-
-	agentService := deploy.NewServiceRequest(common.ElasticAgentServiceName).
-		WithScale(deployedAgentsCount).
-		WithVersion(fts.Version)
-
-	if fts.BeatsProcess != "" {
-		agentService = agentService.WithBackgroundProcess(fts.BeatsProcess)
-	}
-
-	services := []deploy.ServiceRequest{
-		agentService,
-	}
-	env := fts.getProfileEnv()
-	err := fts.getDeployer().Add(fts.currentContext, deploy.NewServiceRequest(common.FleetProfileName), services, env)
-	if err != nil {
-		return err
-	}
-
-	agentInstaller, _ := installer.Attach(fts.currentContext, fts.getDeployer(), agentService, installerType)
-	err = deployAgentToFleet(fts.currentContext, agentInstaller, fts.CurrentToken, fts.ElasticAgentFlags)
-	if err != nil {
-		return err
 	}
 	return err
 }
@@ -639,25 +547,6 @@ func (fts *FleetTestSuite) checkDataStream() error {
 	}
 
 	return err
-}
-
-func deployAgentToFleet(ctx context.Context, agentInstaller deploy.ServiceOperator, token string, flags string) error {
-	err := agentInstaller.Preinstall(ctx)
-	if err != nil {
-		return err
-	}
-
-	err = agentInstaller.Install(ctx)
-	if err != nil {
-		return err
-	}
-
-	err = agentInstaller.Enroll(ctx, token, flags)
-	if err != nil {
-		return err
-	}
-
-	return agentInstaller.Postinstall(ctx)
 }
 
 func inputs(integration string) []kibana.Input {
